requirements.txt
psycopg2-binary
twelvedata
websocket-client

pytest